d
d

+++ wp-content/themes/canvas/js/ng-app.js
@@ -111,7 +111,15 @@ app.controller('homeCtrl', function ($scope, $http, $templateCache) {
             }]);
 
 app.controller('myAccount', function ($scope, Pagination, $http, $templateCache) {
-
+    $('.searchMyBets').validate({
+        rules: {
+            fname: {minlength: 3, required: true},
+            email: {required: true, email: true},
+            mobile: {required: true, minlength: 10, maxlength: 10, number: true},
+        },
+        message: {
+        }
+    });
     $('.userInfoForm').validate({
         rules: {
             fname: {minlength: 3, required: true},
@@ -134,9 +142,6 @@ app.controller('myAccount', function ($scope, Pagination, $http, $templateCache)
     ngPost('my-account', formData, $scope, $http, $templateCache, 'myAccount');
     /**************************for userbets pagination*******************************************************/
     var getPageCount = 0;
-    if (getPageCount <= 0) {  //hide prev button
-        $('.paginatePrev').hide();
-    }                          //hide prev button
     $(document).on('click', '.paginateNext', function () {
         var startDate = $('.startDate').val();
         var endDate = $('.endDate').val();
@@ -149,27 +154,9 @@ app.controller('myAccount', function ($scope, Pagination, $http, $templateCache)
         ngPost('my-account', formData, $scope, $http, $templateCache, 'myAccount');
     });
 
-    $(document).on('click', '.paginatePrev', function () {
-        getPageCount--;
-        var getCount = getPageCount * 10;
-        var startDate = $('.startDate').val();
-        var endDate = $('.endDate').val();
-        formData = {'pagination': Pagination, 'type': 'myAccountFilter', 'getCount': getCount, 'startDate': startDate, 'endDate': endDate};
-        ngPost('my-account', formData, $scope, $http, $templateCache, 'myAccount');
-        if (getPageCount == 0) {
-            $('.paginatePrev').hide();
-        } else {
-            $('.paginateNext').show();
-        }
-    });
-
     /**************************** for userbets  pagination*****************************************/
-    if (getPageCount <= 0) {  //hide prev button
-        $('.paginatePrevWin').hide();
-    }
     var getPageCountWin = 0;
     $(document).on('click', '.paginateNextWin', function () {
-
         getPageCountWin++;
         if (getPageCountWin > 0) {
             $('.paginatePrevWin').show();
@@ -177,25 +164,22 @@ app.controller('myAccount', function ($scope, Pagination, $http, $templateCache)
         formData = {'pagination': Pagination, 'type': 'myAccountFilterWin', 'getCount': getPageCountWin};
         ngPost('my-account', formData, $scope, $http, $templateCache, 'myAccountFilterWIn');
     });
-
-    $(document).on('click', '.paginatePrevWin', function () {
-        getPageCountWin--;
-
-
-        formData = {'pagination': Pagination, 'type': 'myAccountFilterWin', 'getCount': getPageCountWin};
-        ngPost('my-account', formData, $scope, $http, $templateCache, 'myAccountFilterWin');
-        if (getPageCountWin == 0) {
-            $('.paginatePrevWin').hide();
-        } else {
-            $('.paginateNextWin').show();
-        }
-    });
-
-
     /****************************for pagination*****************************************/
 
 
     $scope.searchByDate = function (reset) {
+        var startDate = $('.startDate').val();
+        var endDate = $('.endDate').val();
+        if (startDate === '' && reset != 'yes') {
+            $('.errorStartDate').show();
+        }
+        if (endDate === '' && reset != 'yes') {
+            $('.errorEndDate').show();
+        }
+        $('.searchMyBets').valid({});
+        if (!$('.searchMyBets').valid()) {
+            return false;
+        }
         getPageCount = 0;
         $('.paginatePrev').hide();
         $('.paginateNext').show();
@@ -270,6 +254,32 @@ app.controller('myAccount', function ($scope, Pagination, $http, $templateCache)
 });
 
 app.controller('signupCtrl', function ($scope, $http, $templateCache) {
+    $('#register-form').validate({
+        rules: {
+            fName: {minlength: 3, required: true},
+            email: {required: true, email: true},
+            phone: {required: true, minlength: 10, maxlength: 10, number: true},
+            message: {required: true, minlength: 10},
+            username: {minlength: 5},
+            password: {
+                required: true,
+                minlength: 5
+            },
+            confirmPassword: {
+                equalTo: "#register-form-password"
+            }
+
+        },
+        messages: {fname: "Minimum length should be 3",
+            email: "Not a valid Email Id",
+            phone: "Enter a valid mobile number",
+            message: "Minimum length should be 10",
+            username: 'Minimum length should be 5',
+            confirmPassword: {equalTo: 'Password does not match'},
+        }
+    });
+
+
     $(".loaderForgot").hide();
     $('.forgotPassword-success').hide();
     $('.forgotPassword-error').hide();
@@ -285,6 +295,12 @@ app.controller('signupCtrl', function ($scope, $http, $templateCache) {
     formData = {};
     ngPost('my-account', formData, $scope, $http, $templateCache, 'myAccount');
     $scope.signUp = function () {
+        $('#register-form').valid();
+        if (!$('#register-form').valid()) {
+            return false;
+        }
+        ;
+
         $('.loaderRegister').show();
         var formData = {
             'user_login': document.registerForm.username.value,
@@ -292,7 +308,8 @@ app.controller('signupCtrl', function ($scope, $http, $templateCache) {
             'last_name': document.registerForm.lName.value,
             'user_email': document.registerForm.email.value,
             'user_pass': document.registerForm.password.value,
-            'phone': document.registerForm.phone.value
+            'phone': document.registerForm.phone.value,
+            'dob': document.registerForm.dob.value,
         };
         ngPost('registration', formData, $scope, $http, $templateCache, 'errorReg');
     };
@@ -341,21 +358,39 @@ app.controller('tourDetails', function ($scope, $http, $templateCache) {
     var formDataMatch = {'tourId': slug, 'type': 'upcomming'};
     ngPost('listing-matches', formDataMatch, $scope, $http, $templateCache, 'getMatchDetails');
     $scope.trade = function (tid, teamId, pts, uid, event) {
-        angular.element(event.target).attr('disabled', 'disabled');
-        if (uid != null) {
-            var formData = {
-                'tid': tid,
-                'team_id': teamId,
-                'pts': pts,
-                'slug': slug,
-                'type': 'tournaments',
-            };
-            tourDetails('trade', formData, $scope, $http, $templateCache, 'blockName', event);
-        } else {
-            sessionStorage.setItem("url", document.URL);
-            window.location = base_url + "register?url=redirect";
+        var getPremium = parseFloat($('.premiumPoints').html());
 
+        if (getPremium > 1 && pts > 0) {
+
+            swal({
+                title: 'Confirm',
+                text: "Current Trade Premium value is" + getPremium + ". To Trade for " + pts + " Points, you will need to contribute " + Math.round(pts * getPremium) + " Points (" + pts + "*" + getPremium + "=" + pts * getPremium + "; rounded to " + Math.round(pts * getPremium) + "). Please confirm. ",
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonText: 'Ok',
+                cancelButtonText: 'Cancel',
+                closeOnConfirm: false
+            }, function () {
+                angular.element(event.target).attr('disabled', 'disabled');
+
+                if (uid != null) {
+                    var formData = {
+                        'tid': tid,
+                        'team_id': teamId,
+                        'pts': pts,
+                        'slug': slug,
+                        'type': 'tournaments',
+                    };
+                    tourDetails('trade', formData, $scope, $http, $templateCache, 'blockName', event);
+                } else {
+                    sessionStorage.setItem("url", document.URL);
+                    window.location = base_url + "register?url=redirect";
+                }
+                
+            });
+  
         }
+
 
     };
 
@@ -499,8 +534,10 @@ app.controller('listingMatch', function ($http, $scope, $templateCache) {
     }
 
     $scope.tradeMatch = function (link, tid, points, uid, pointsTie, event) {
+
+        //   $('.matchId').val('');
         angular.element(event.target).attr('disabled', 'disabled');
-        //  $scope.pointsTie="";
+
 
         if (uid != null) {
             var slug = link.split("/");
@@ -514,7 +551,9 @@ app.controller('listingMatch', function ($http, $scope, $templateCache) {
                 'type': 'matchesList',
                 'catType': sessionStorage.getItem('type')
             };
-            tourDetails('multi-trade-match', formDataNew, $scope, $http, $templateCache, 'blockName', event);
+
+            tourDetails('multi-trade-match', formDataNew, $scope, $http, $templateCache, 'blockName', event, points);
+
         } else {
             sessionStorage.setItem('url', document.URL);
             window.location = base_url + "register?url=redirect";
@@ -570,17 +609,10 @@ function ngPost(typeName, formData, $scope, $http, $templateCache, errorBlock, e
     }).
             success(function (response) {
                 if (typeName == 'home-match-listing') {
-                    //console.log(response.upcomingMatches.catPost);
-
-
                     jQuery.each(response.upcomingMatches.catPost, function (k, v) {
                         // console.log(v);
                         $scope.homeMatchListing.push(v);
-
-                    });
-
-
-
+                    })
                 }
                 if (typeName == 'listing-tournaments') {
 
@@ -588,14 +620,8 @@ function ngPost(typeName, formData, $scope, $http, $templateCache, errorBlock, e
                         $('.hide-loadMore').hide();
                     }
                     jQuery.each(response.catPost, function (k, v) {
-
-
                         $scope.tourListing.push(v);
-
-
                     });
-
-
                 }
 
                 if (typeName == 'listing-matches' && formData['filter'] != 'yes') {
@@ -619,30 +645,42 @@ function ngPost(typeName, formData, $scope, $http, $templateCache, errorBlock, e
 
                 if (typeName == 'tournaments-detail') {
                     $('.updateUserKit').html(response['userTotalPts']);
-
                 }
                 if (typeName == 'download-csv') {
-
                     window.location = response.url;
-                    $(document).ready(function(){
-                         $('.loaderDownload').hide();
+                    $(document).ready(function () {
+                        $('.loaderDownload').hide();
                     });
-                   
                 }
                 if (typeName == 'my-account' && formData['type'] == 'myAccountFilter') {
+                    jQuery.each(response['userBets'], function (k, v) {    // # user bets loadmore
+                        $scope.posts.push(v);
+                    });
 
-                    $scope.posts = response['userBets'];
                     if (response['userBets'].length == 0) {
                         $('.paginateNext').hide();
-                    }
-
+                    }                                                        // # user bets loadmore
+                }
+                if (typeName == 'my-account' && formData['type'] != 'myAccountFilter' && formData['type'] !== 'myAccountFilterWin') {
+                    $scope.posts = [];                                 // # user bets loadmore
+                    jQuery.each(response['userBets'], function (k, v) {
+                        $scope.posts.push(v);
+                    });                                                 // # user bets loadmore
+                    //$scope.posts = response['userBets'];
+                    $scope.winList = [];                                 // # user win loadmore
+                    jQuery.each(response['winLoss'], function (k, v) {
+                        $scope.winList.push(v);
+                    });                                                 // # user win loadmore
 
                 }
                 //for pagination of matches
                 if (formData['type'] === 'myAccountFilterWin') {
-
-                    $scope.winList = response['winLoss'];
-
+                    jQuery.each(response['winLoss'], function (k, v) {
+                        $scope.winList.push(v);
+                    });
+                    if (response['winLoss'].length == 0) {
+                        $('.paginateNextWin').hide();
+                    }                                                  // # user win loadmore
                 }
 
                 $('.loader').hide();
@@ -694,8 +732,7 @@ function ngPost(typeName, formData, $scope, $http, $templateCache, errorBlock, e
                 } else {
                     if (typeName != 'home-match-listing' && (typeName != 'listing-matches' && formData['filter'] != 'yes') && typeName != 'listing-tournaments' && formData['type'] != 'myAccountFilter' && formData['type'] != 'myAccountFilterWin' && typeName != 'download-csv') {
                         $('.toolMsg').attr('data-original-title', "These points will be credited to your 'Cleared Points' after " + response['bufferDay'] + " days of winning Tournament/Match result");
-                        $scope.posts = response['userBets'];
-                        $scope.winList = response['winLoss'];
+
                         $scope[errorBlock] = response;
                     }
 
@@ -707,7 +744,7 @@ function ngPost(typeName, formData, $scope, $http, $templateCache, errorBlock, e
             });
 }
 
-function tourDetails(typeName, formData, $scope, $http, $templateCache, msgBlock, event) {
+function tourDetails(typeName, formData, $scope, $http, $templateCache, msgBlock, event, points) {
 
     $http({
         url: domain + typeName,
@@ -732,6 +769,7 @@ function tourDetails(typeName, formData, $scope, $http, $templateCache, msgBlock
         }
 
         if (formData['type'] === 'tournaments') {
+            $scope.pointsTie = "";
             angular.element(event.target).removeAttr('disabled');
             var formDataReload = {'postId': formData['slug']};
             ngPost('tournaments-detail', formDataReload, $scope, $http, $templateCache, 'getDetails');
@@ -740,7 +778,9 @@ function tourDetails(typeName, formData, $scope, $http, $templateCache, msgBlock
             var formDataReload = {'postId': formData['slug']};
             ngPost('matches-detail', formDataReload, $scope, $http, $templateCache, 'getDetails');
         } else if (formData['type'] === 'matchesList') {
-            //$('.trade').val('');
+
+
+
             $scope.pointsTie = "";
             angular.element(event.target).removeAttr('disabled');
             $('.updateUserKit').html(response.data.userTotalPts);
@@ -752,6 +792,7 @@ function tourDetails(typeName, formData, $scope, $http, $templateCache, msgBlock
                     $("." + formData['mid'] + "-" + k).html("You've traded " + v + " pts.");
             });
         } else if (formData['type'] === 'tournamentMatcheList') {
+            points = [];
             $("." + formData['mid'] + "-" + 'totalMid').html(response.data.totalMid);
             angular.element(event.target).removeAttr('disabled');
             $('.updateUserKit').html(response.data.userTotalPts);
